% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{absorb.reorg}
\alias{absorb.reorg}
\title{Absorb Fixed Effects and Reorder Data}
\usage{
absorb.reorg(data, x.names, cluster.id, weights, fe.id, crve, y.name = NULL)
}
\arguments{
\item{data}{Data frame or matrix containing all outcomes, covariates, and group ids}

\item{x.names}{Vector of covariate names}

\item{cluster.id}{Name of cluster id variable}

\item{weights}{Name of weight variable, if applicable}

\item{fe.id}{Name of fixed effect group id variable (defaults to cluster.id)}

\item{crve}{Type of cluster-robust variance estimator (CR0, CR1, CR2, CR3)}
}
\value{
A list containing: Xdw.h (covariates by cluster),
XTX.inv (inverse squared design matrix),
A.h (adjustment matrices by cluster),
Ydw.h (outcomes by cluster, if applicable)
}
\description{
This function reorganizes the data by cluster and fixed effect,
absorbing group fixed effects and applying weights.
}
