% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{adjustment.matrices}
\alias{adjustment.matrices}
\title{Adjustment Matrices}
\usage{
adjustment.matrices(xg, xtx, delta = 0.5)
}
\arguments{
\item{xg}{List of cluster-specific submatrices of design matrix}

\item{xtx}{Inverse of design matrix squared}

\item{delta}{Power to raise (I_g - H_gg), by default delta=0.5 for CR2}
}
\value{
A vector of eigenvalues
}
\description{
This function gets adjustment matrices for the
cluster-robust variance estimators CR2 and CR3
from Bell and McCaffrey (2002).
}
