% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{bm.test.lambdas}
\alias{bm.test.lambdas}
\title{Eigenvalues for BM Test}
\usage{
bm.test.lambdas(xg, xtx, ag, c.0)
}
\arguments{
\item{xg}{xg List of cluster-specific submatrices of design matrix}

\item{xtx}{Inverse of design matrix squared}

\item{ag}{List of adjustment matrices}

\item{c.0}{Linear hypothesis vector}
}
\value{
A vector of eigenvalues
}
\description{
This function gets the eigenvalues for a test based on
Bell and McCaffrey (2002).
}
