% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{conf.interval.meis}
\alias{conf.interval.meis}
\title{Calculate Confidence Interval}
\usage{
conf.interval.meis(
  data,
  y.name,
  x.names,
  cluster.id,
  alpha = 0.05,
  weights = NULL,
  c0 = NULL,
  test.var = NULL,
  fe.id = cluster.id,
  crve.type = "CR1"
)
}
\arguments{
\item{data}{Data frame or matrix containing all outcomes, covariates, and group ids}

\item{y.name}{Name of dependent variable}

\item{x.names}{Vector of covariate names}

\item{cluster.id}{Name of cluster id variable}

\item{alpha}{Size of test associated with the confidence interval (e.g. a test of size 0.05 is associated with a 95\% confidence interval)}

\item{weights}{Name of weight variable, if applicable}

\item{c0}{Linear hypothesis vector, either c0 or test.var must be supplied}

\item{test.var}{Index of variable to test, either c0 or test.var must be supplied}

\item{fe.id}{Name of fixed effect group id variable (defaults to cluster.id)}

\item{crve.type}{Type of cluster-robust variance estimator (CR0, CR1, CR2, CR3)}
}
\value{
numeric, the lower and upper bound of the confidence interval
}
\description{
This function calculates a confidence interval for the
coefficient (or linear combination of coefficients) defined
by test.var (or c0) according to Meiselman (2021).
}
