% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{matrix_power}
\alias{matrix_power}
\title{Matrix Exponent}
\usage{
matrix_power(x, p, symmetric = TRUE, tol = -12)
}
\arguments{
\item{x}{A square matrix}

\item{p}{Linear hypothesis vector}

\item{symmetric}{Indicates whether x is a symmetric matrix (if true, improves performance slightly)}

\item{tol}{Error tolerance for eigenvalues close to zero assumed to be zero}
}
\value{
A vector of eigenvalues
}
\description{
This function raises a matrix to an exponent using
the matrix's eigendecomposition.
}
