% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{meis.test.lambdas.1}
\alias{meis.test.lambdas.1}
\title{Eigenvalues for Meiselman Test}
\usage{
meis.test.lambdas.1(xg, xtx, ag, c.0)
}
\arguments{
\item{xg}{List of cluster-specific submatrices of design matrix}

\item{xtx}{Inverse of design matrix squared}

\item{c.0}{Linear hypothesis vector}
}
\value{
A list containing two principal submatrices of the key matrix
}
\description{
This function gets an intermediate thing for the eigenvalues
for a test based on Meiselman (2021).
}
