% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{meis.test.lambdas.2}
\alias{meis.test.lambdas.2}
\title{Eigenvalues for Meiselman Test}
\usage{
meis.test.lambdas.2(m1, c.0, q)
}
\arguments{
\item{m1}{A list containing two principal submatrices of the key matrix}

\item{c.0}{Linear hypothesis vector}

\item{q}{Desired quantile}
}
\value{
A vector of eigenvalues
}
\description{
This function gets the eigenvalues for a test based on
Meiselman (2021).
}
