% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{meis.test.lambdas.slow}
\alias{meis.test.lambdas.slow}
\title{Eigenvalues for Meiselman Test}
\usage{
meis.test.lambdas.slow(
  xtx,
  ag,
  c.0,
  t.0,
  omg,
  xs,
  ihg,
  ind.hi,
  ind.lo,
  D0 = NULL
)
}
\arguments{
\item{xtx}{Inverse of design matrix squared}

\item{c.0}{Linear hypothesis vector}

\item{xg}{List of design matrix, separated by cluster}
}
\value{
A vector of eigenvalues
}
\description{
This function gets the eigenvalues for a test based on
Meiselman (2021).
}
