% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{p.meis}
\alias{p.meis}
\title{P-Value for Meiselman Test}
\usage{
p.meis(m1, c.0, q)
}
\arguments{
\item{m1}{A list containing two principal submatrices of the key matrix}

\item{c.0}{Linear hypothesis vector}

\item{q}{Desired quantile}
}
\value{
Probability that abs(test statistic) is greater than q
}
\description{
This function gets the p-value for a test based on
Meiselman (2021).
}
