% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{p.value.meis}
\alias{p.value.meis}
\title{Calculate P-Value}
\usage{
p.value.meis(
  t0,
  data,
  x.names,
  cluster.id,
  weights = NULL,
  c0 = NULL,
  test.var = NULL,
  fe.id = cluster.id,
  crve.type = "CR1"
)
}
\arguments{
\item{t0}{The test statistic}

\item{data}{Data frame or matrix containing all outcomes, covariates, and group ids}

\item{x.names}{Vector of covariate names}

\item{cluster.id}{Name of cluster id variable}

\item{weights}{Name of weight variable, if applicable}

\item{c0}{Linear hypothesis vector, either c0 or test.var must be supplied}

\item{test.var}{Index of variable to test, either c0 or test.var must be supplied}

\item{fe.id}{Name of fixed effect group id variable (defaults to cluster.id)}

\item{crve.type}{Type of cluster-robust variance estimator (CR0, CR1, CR2, CR3)}
}
\value{
numeric, the probability that a test statistic with a magnitude of abs(t0) or greater is observed
}
\description{
This function calculates a p-value for the test statistic t0
according to Meiselman (2021).
}
