% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{q.meis}
\alias{q.meis}
\title{Critical Value for Meiselman Test}
\usage{
q.meis(m1, c.0, alpha = 0.05)
}
\arguments{
\item{m1}{A list containing two principal submatrices of the key matrix}

\item{c.0}{Linear hypothesis vector}

\item{q}{Desired quantile}
}
\value{
Critical value for a test with size alpha
}
\description{
This function gives the critical value for a
test based on Meiselman (2021).
}
