% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\name{test.params}
\alias{test.params}
\title{Test Parameters}
\usage{
test.params(
  data,
  x.names,
  cluster.id,
  weights = NULL,
  c0 = NULL,
  test.var = NULL,
  fe.id = cluster.id,
  calling.function = "unknown"
)
}
\arguments{
\item{data}{Data frame or matrix containing all outcomes, covariates, and group ids}

\item{x.names}{Vector of covariate names}

\item{cluster.id}{Name of cluster id variable}

\item{weights}{Name of weight variable, if applicable}

\item{c0}{Linear hypothesis vector, either c0 or test.var must be supplied}

\item{test.var}{Index of variable to test, either c0 or test.var must be supplied}

\item{fe.id}{Name of fixed effect group id variable (defaults to cluster.id)}

\item{calling.function}{Function whose parameters are being checked}
}
\value{
TRUE if no errors, otherwise does not return
}
\description{
This function tests the validity of several basic
model parameters.
}
